+++ js/PreConfig.js
@@ -8,6 +8,7 @@ window.EXPORT_URL = 'REPLACE_WITH_YOUR_IMAGE_SERVER';
 window.PLANT_URL = window.top.siyuan?.config.editor.plantUMLServePath ?? 'https://www.plantuml.com/plantuml/svg/~1';
 window.DRAWIO_BASE_URL = `${window.location.origin}${window.location.pathname}`.replace(/\/$/, '')
 window.DRAWIO_VIEWER_URL = `${window.DRAWIO_BASE_URL}/js/viewer.min.js`
+window.ALLOW_CUSTOM_PLUGINS = true;
 /* 👆 SIYUAN 👆 */
 window.DRAWIO_LIGHTBOX_URL = null; // Replace with your lightbox URL, eg. https://www.example.com
 window.DRAW_MATH_URL = 'math/es5';